requirements.txt
# Django and core dependencies
Django==5.2.10
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Message Queue (RabbitMQ)
pika==1.3.2

# APIs and HTTP
requests==2.31.0
stripe==5.4.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
Pillow==10.1.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Async task queue (for future use)
celery==5.3.4
redis==5.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==20.1.0

# Code quality
black==23.12.0
flake8==6.1.0
isort==5.13.2
pylint==3.0.3
